# international_cuisine_backend/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully"},"409":{"description":"Email or username already exists"}},"summary":"Register a new user","tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful, returns JWT token"},"401":{"description":"Invalid credentials"}},"summary":"Login with email and password","tags":["Auth"]}},"/api/v1/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"Current user profile"},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Auth"]}},"/api/v1/users/profile":{"patch":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"Profile updated successfully"}},"security":[{"bearer":[]}],"summary":"Update user profile","tags":["Users"]}},"/api/v1/users/achievements":{"get":{"operationId":"UsersController_getAchievements","parameters":[],"responses":{"200":{"description":"User achievements"}},"security":[{"bearer":[]}],"summary":"Get user achievements (badges, unlocked recipes)","tags":["Users"]}},"/api/v1/users/level":{"get":{"operationId":"UsersController_getLevelInfo","parameters":[],"responses":{"200":{"description":"Level info with XP progress"}},"security":[{"bearer":[]}],"summary":"Get user level info and XP progress","tags":["Users"]}},"/api/v1/recipes":{"get":{"operationId":"RecipesController_findAll","parameters":[{"name":"search","required":false,"in":"query","schema":{"example":"sushi","type":"string"}},{"name":"country","required":false,"in":"query","schema":{"example":"Japan","type":"string"}},{"name":"continent","required":false,"in":"query","schema":{"example":"Asia","type":"string"}},{"name":"diet","required":false,"in":"query","schema":{"example":"vegetarian","type":"string"}},{"name":"difficulty","required":false,"in":"query","schema":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]}},{"name":"maxTime","required":false,"in":"query","schema":{"example":30,"type":"number"}},{"name":"maxIngredients","required":false,"in":"query","schema":{"example":6,"type":"number"}},{"name":"maxCalories","required":false,"in":"query","schema":{"example":500,"type":"number"}},{"name":"cookingMethod","required":false,"in":"query","schema":{"example":"fry","type":"string"}},{"name":"showHidden","required":false,"in":"query","schema":{"example":false,"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"example":20,"type":"number"}}],"responses":{"200":{"description":"Paginated recipe list"}},"summary":"List recipes with search and filters","tags":["Recipes"]},"post":{"operationId":"RecipesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"201":{"description":"Recipe created successfully"}},"security":[{"bearer":[]}],"summary":"Create a new recipe","tags":["Recipes"]}},"/api/v1/recipes/{id}":{"patch":{"operationId":"RecipesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeDto"}}}},"responses":{"200":{"description":"Recipe updated successfully"},"403":{"description":"Forbidden: You can only update your own recipes"}},"security":[{"bearer":[]}],"summary":"Update your own recipe","tags":["Recipes"]},"delete":{"operationId":"RecipesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Recipe deleted successfully"},"403":{"description":"Forbidden: You can only delete your own recipes"}},"security":[{"bearer":[]}],"summary":"Delete your own recipe","tags":["Recipes"]},"get":{"operationId":"RecipesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Recipe with ingredients and steps"},"404":{"description":"Recipe not found"}},"summary":"Get recipe detail (auto-increments view count)","tags":["Recipes"]}},"/api/v1/recipes/countries":{"get":{"operationId":"RecipesController_getCountries","parameters":[],"responses":{"200":{"description":"Countries grouped by continent"}},"summary":"Get available countries grouped by continent","tags":["Recipes"]}},"/api/v1/recipes/{id}/scale":{"get":{"operationId":"RecipesController_getScaled","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"servings","required":true,"in":"query","schema":{"example":4,"type":"number"}}],"responses":{"200":{"description":"Recipe with scaled ingredient quantities"}},"summary":"Get recipe with scaled servings","tags":["Recipes"]}},"/api/v1/recipes/{id}/complete":{"post":{"operationId":"RecipesController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Completion recorded, XP and level info returned"}},"security":[{"bearer":[]}],"summary":"Mark recipe as completed (earn XP, check level up and unlocks)","tags":["Recipes"]}},"/api/v1/recipes/{id}/unlock":{"post":{"operationId":"RecipesController_unlock","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Unlock result with condition status"}},"security":[{"bearer":[]}],"summary":"Attempt to unlock a hidden recipe","tags":["Recipes"]}},"/api/v1/favorites":{"get":{"operationId":"FavoritesController_findAll","parameters":[{"name":"category","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"List of favorites"}},"security":[{"bearer":[]}],"summary":"List user favorites, optionally by category","tags":["Favorites"]},"post":{"operationId":"FavoritesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFavoriteDto"}}}},"responses":{"201":{"description":"Added to favorites"},"409":{"description":"Already in favorites"}},"security":[{"bearer":[]}],"summary":"Add a recipe to favorites","tags":["Favorites"]}},"/api/v1/favorites/{id}":{"delete":{"operationId":"FavoritesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Removed from favorites"},"404":{"description":"Favorite not found"}},"security":[{"bearer":[]}],"summary":"Remove a recipe from favorites","tags":["Favorites"]}},"/api/v1/favorites/categories":{"get":{"operationId":"FavoritesController_getCategories","parameters":[],"responses":{"200":{"description":"List of category names"}},"security":[{"bearer":[]}],"summary":"List custom favorite categories","tags":["Favorites"]}},"/api/v1/shopping-lists":{"get":{"operationId":"ShoppingListsController_findAll","parameters":[],"responses":{"200":{"description":"User shopping lists"}},"security":[{"bearer":[]}],"summary":"List all shopping lists","tags":["Shopping Lists"]},"post":{"operationId":"ShoppingListsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateShoppingListDto"}}}},"responses":{"201":{"description":"Shopping list created with merged ingredients"}},"security":[{"bearer":[]}],"summary":"Create shopping list from recipe IDs with servings (auto-merges ingredients)","tags":["Shopping Lists"]}},"/api/v1/shopping-lists/{id}":{"get":{"operationId":"ShoppingListsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Shopping list with items"}},"security":[{"bearer":[]}],"summary":"Get shopping list detail","tags":["Shopping Lists"]},"delete":{"operationId":"ShoppingListsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Shopping list deleted"}},"security":[{"bearer":[]}],"summary":"Delete a shopping list","tags":["Shopping Lists"]}},"/api/v1/shopping-lists/{id}/items/{itemId}/toggle":{"patch":{"operationId":"ShoppingListsController_toggleItem","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"itemId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Item toggled"}},"security":[{"bearer":[]}],"summary":"Toggle item checked status","tags":["Shopping Lists"]}},"/api/v1/shopping-lists/{id}/export":{"get":{"operationId":"ShoppingListsController_exportCsv","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"CSV file download"}},"security":[{"bearer":[]}],"summary":"Export shopping list as CSV","tags":["Shopping Lists"]}},"/api/v1/recipes/{recipeId}/ratings":{"get":{"operationId":"RatingsController_findByRecipe","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"List of ratings with user info"}},"summary":"Get ratings for a recipe","tags":["Ratings"]},"post":{"operationId":"RatingsController_create","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRatingDto"}}}},"responses":{"201":{"description":"Rating created"},"409":{"description":"Already rated"}},"security":[{"bearer":[]}],"summary":"Rate and comment on a recipe (1-5 stars)","tags":["Ratings"]}},"/api/v1/recipes/{recipeId}/ratings/{id}":{"patch":{"operationId":"RatingsController_update","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRatingDto"}}}},"responses":{"200":{"description":"Rating updated"}},"security":[{"bearer":[]}],"summary":"Update own rating","tags":["Ratings"]},"delete":{"operationId":"RatingsController_remove","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Rating deleted"}},"security":[{"bearer":[]}],"summary":"Delete own rating","tags":["Ratings"]}},"/api/v1/analytics/events":{"post":{"operationId":"AnalyticsController_trackEvent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"Event recorded"}},"security":[{"bearer":[]}],"summary":"Track an analytics event (view, start, complete, favorite)","tags":["Analytics"]}},"/api/v1/analytics/stats":{"get":{"operationId":"AnalyticsController_getStats","parameters":[],"responses":{"200":{"description":"User stats overview"}},"security":[{"bearer":[]}],"summary":"Get user analytics stats","tags":["Analytics"]}},"/api/v1/admin/recipes":{"post":{"operationId":"AdminController_createRecipe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"201":{"description":"Recipe created"},"403":{"description":"Admin access required"}},"security":[{"bearer":[]}],"summary":"[Admin] Create a new recipe with ingredients and steps","tags":["Admin"]}},"/api/v1/admin/recipes/{id}":{"patch":{"operationId":"AdminController_updateRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeDto"}}}},"responses":{"200":{"description":"Recipe updated"}},"security":[{"bearer":[]}],"summary":"[Admin] Update a recipe","tags":["Admin"]},"delete":{"operationId":"AdminController_deleteRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Recipe deleted"}},"security":[{"bearer":[]}],"summary":"[Admin] Delete a recipe","tags":["Admin"]}},"/api/v1/admin/recipes/{id}/unlock-conditions":{"post":{"operationId":"AdminController_setUnlockConditions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"Unlock conditions set"}},"security":[{"bearer":[]}],"summary":"[Admin] Set unlock conditions for a hidden recipe","tags":["Admin"]}},"/api/v1/admin/analytics/overview":{"get":{"operationId":"AdminController_getAnalyticsOverview","parameters":[],"responses":{"200":{"description":"Dashboard stats with top recipes"}},"security":[{"bearer":[]}],"summary":"[Admin] Get analytics dashboard overview","tags":["Admin"]}},"/api/v1/admin/users":{"get":{"operationId":"AdminController_listUsers","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Paginated user list"}},"security":[{"bearer":[]}],"summary":"[Admin] List all users with stats","tags":["Admin"]}}},"info":{"title":"International Cuisine API","description":"Backend API for the International Cuisine App - recipe browsing, user management, gamification, shopping lists, and admin content management.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"securePassword123"},"username":{"type":"string","example":"chef_master"}},"required":["email","password","username"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"securePassword123"}},"required":["email","password"]},"UpdateProfileDto":{"type":"object","properties":{"username":{"type":"string","example":"new_username"},"avatarUrl":{"type":"string","example":"https://example.com/avatar.jpg"},"location":{"type":"string","example":"Taipei, Taiwan"},"dietaryPreferences":{"example":["vegetarian","no-beef"],"description":"Dietary preferences array","type":"array","items":{"type":"string"}}}},"IngredientDto":{"type":"object","properties":{"name":{"type":"string","example":"Soy Sauce"},"qtyPerServing":{"type":"number","example":15},"unit":{"type":"string","example":"ml"},"optional":{"type":"boolean","example":false},"notes":{"type":"string","example":"Use dark soy sauce for better color"},"gramEquivalent":{"type":"number","example":15},"category":{"type":"string","enum":["VEGETABLE","MEAT","DRY_GOODS","SEASONING","REFRIGERATED","OTHER"]},"scaleStrategy":{"type":"string","enum":["LINEAR","CLAMPED","MIN_UNIT"]},"kcalPer100g":{"type":"number","example":60}},"required":["name","qtyPerServing","unit"]},"StepDto":{"type":"object","properties":{"order":{"type":"number","example":1},"description":{"type":"string","example":"Mix all the sauce ingredients together"},"defaultTimerSeconds":{"type":"number","example":30},"mediaUrl":{"type":"string","example":"https://storage.example.com/step1.jpg"}},"required":["order","description"]},"UnlockConditionDto":{"type":"object","properties":{"conditionType":{"type":"string","example":"completion_count"},"conditionData":{"type":"object","example":{"count":10}},"description":{"type":"string","example":"Complete any 10 recipes to unlock"}},"required":["conditionType","conditionData","description"]},"CreateRecipeDto":{"type":"object","properties":{"title":{"type":"string","example":"Japanese Tonkatsu Ramen"},"altTitles":{"example":["Pork Bone Ramen","Tonkotsu"],"type":"array","items":{"type":"string"}},"country":{"type":"string","example":"Japan"},"region":{"type":"string","example":"Asia"},"city":{"type":"string","example":"Tokyo"},"cuisineTags":{"example":["ramen","soup","pork"],"type":"array","items":{"type":"string"}},"isHidden":{"type":"boolean","example":false},"prepTimeMinutes":{"type":"number","example":30},"cookTimeMinutes":{"type":"number","example":120},"totalTimeMinutes":{"type":"number","example":150},"difficulty":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"servingsBase":{"type":"number","example":1},"caloriesPerServing":{"type":"number","example":550},"culturalNote":{"type":"string","example":"Tonkotsu ramen originated in Fukuoka, Japan, in the 1940s."},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/IngredientDto"}},"steps":{"type":"array","items":{"$ref":"#/components/schemas/StepDto"}},"unlockConditions":{"type":"array","items":{"$ref":"#/components/schemas/UnlockConditionDto"}}},"required":["title","country","prepTimeMinutes","cookTimeMinutes","totalTimeMinutes","difficulty","ingredients","steps"]},"UpdateRecipeDto":{"type":"object","properties":{"title":{"type":"string","example":"Updated Recipe Title"},"country":{"type":"string","example":"Japan"},"region":{"type":"string","example":"Asia"},"isHidden":{"type":"boolean","example":false},"prepTimeMinutes":{"type":"number","example":30},"cookTimeMinutes":{"type":"number","example":120},"totalTimeMinutes":{"type":"number","example":150},"difficulty":{"type":"string","enum":["BEGINNER","INTERMEDIATE","ADVANCED"]},"caloriesPerServing":{"type":"number","example":550},"culturalNote":{"type":"string"}}},"CreateFavoriteDto":{"type":"object","properties":{"recipeId":{"type":"string","example":"recipe-uuid-here"},"categoryName":{"type":"string","example":"dinner","description":"Custom category name for organizing favorites"}},"required":["recipeId"]},"RecipeServingPair":{"type":"object","properties":{"recipeId":{"type":"string","example":"recipe-uuid-here"},"servings":{"type":"number","example":2}},"required":["recipeId","servings"]},"CreateShoppingListDto":{"type":"object","properties":{"title":{"type":"string","example":"Weekend Cooking"},"recipes":{"description":"Array of recipe IDs with desired servings","type":"array","items":{"$ref":"#/components/schemas/RecipeServingPair"}}},"required":["recipes"]},"CreateRatingDto":{"type":"object","properties":{"score":{"type":"number","example":5,"description":"Rating score from 1 to 5"},"comment":{"type":"string","example":"Delicious recipe! Very easy to follow."}},"required":["score"]},"UpdateRatingDto":{"type":"object","properties":{"score":{"type":"number","example":4},"comment":{"type":"string","example":"Updated: Still great but added more spice!"}}},"CreateEventDto":{"type":"object","properties":{"recipeId":{"type":"string","example":"recipe-uuid-here"},"eventType":{"type":"string","enum":["VIEW","START","COMPLETE","FAVORITE"]},"meta":{"type":"object","example":{"source":"home_page","duration_seconds":120}}},"required":["eventType"]}}}}